Correções DB

+++ src/main/java/br/com/cdweb/persistence/domain/AcaoPredefinicao.java
@@ -5,7 +5,6 @@ import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
@@ -26,15 +25,15 @@ public class AcaoPredefinicao extends ComunEntidades implements Serializable{
 	@Id    
     @GeneratedValue(generator="acao_predefinicao_id_acao_predefinicao_seq")
     @Column(name = "id_acao_predeficao", updatable=false)
-	private int idAcaoPredeficao;
+	private long idAcaoPredeficao;
 	
 	public AcaoPredefinicao() {
 	}
 	
-	public int getIdAcaoPredeficao() {
+	public long getIdAcaoPredeficao() {
 		return idAcaoPredeficao;
 	}
-	public void setIdAcaoPredeficao(int idAcaoPredeficao) {
+	public void setIdAcaoPredeficao(long idAcaoPredeficao) {
 		this.idAcaoPredeficao = idAcaoPredeficao;
 	}
